Break out dash/open_weather.py
