+++ dash/board.py
@@ -11,28 +11,11 @@ from fastapi.staticfiles import StaticFiles
 app = FastAPI()
 
 
-OPEN_WEATHER_BASE_URL = "https://api.openweathermap.org/data/2.5/forecast"
 OPEN_METEO_BASE_URL = "https://api.open-meteo.com/v1/forecast"
 WEATHER_API_BASE_URL = "http://api.weatherapi.com/v1/current.json"
 
-OPENWEATHER_API_KEY = os.getenv("OPENWEATHER_API_KEY")
 WEATHER_API_KEY = os.getenv("WEATHER_API_KEY")
-# assert len(OPENWEATHER_API_KEY) == 32
 # assert len(WEATHER_API_KEY) == 31
-
-
-def open_weather_fetch_forecast(zip_code: str) -> dict[str, float]:
-    assert len(zip_code) == 5, zip_code
-    params = {
-        "zip": zip_code,
-        "appid": OPENWEATHER_API_KEY,
-        "units": "imperial",  # or 'metric' for Celsius
-    }
-    response = httpx.get(OPEN_WEATHER_BASE_URL, params=params)
-    response.raise_for_status()
-    payload = dict(response.json())
-    hourly = list(payload["list"])
-    return {d["dt_txt"]: round(d["main"]["temp"], 1) for d in hourly}
 
 
 def open_meteo_fetch_forecast(zip_code: str, country_code: str = "US") -> dict:  # type: ignore
